requirements.txt
baidu_aip==4.16.7
beautifulsoup4==4.11.1
fontTools==4.38.0
lxml==4.9.1
mysql_connector_repackaged==0.3.1
numpy==1.23.1
opencv_python==4.6.0.66
pandas==1.4.3
parsel==1.6.0
Pillow==9.3.0
requests==2.28.1
Scrapy==2.6.2
selenium==4.5.0
